feat: 🎸 move error_handler to http_response
