+++ src/components/ambulance/services.rs
@@ -7,7 +7,7 @@ use crate::entity::sea_orm_active_enums::{
 };
 
 use crate::entity::hospital;
-use crate::error_handler::CustomError;
+use crate::http_response::error_handler::CustomError;
 use crate::shared::{PaginatedResponse, PaginationInfo};
 use crate::utils::helpers::{check_if_is_duplicate_key_from_data_base, generate_ic, now_time};
