+++ src/components/appointment/routes.rs
@@ -1,6 +1,6 @@
 use super::services::AppointmentService;
 use crate::entity::appointment::AppointmentRequestBody;
-use crate::error_handler::CustomError;
+use crate::http_response::error_handler::CustomError;
 use crate::http_response::{HttpCodeW, http_response_builder};
 use actix_web::{HttpResponse, post, web};
 use sea_orm::DatabaseConnection;
