+++ src/components/department/services.rs
@@ -1,7 +1,7 @@
 use crate::components::hospital::HospitalService;
 use crate::entity::department::{ActiveModel, Column, DepartmentRequestBody, Entity, Model};
 use crate::entity::sea_orm_active_enums::DepartmentNameEnum;
-use crate::error_handler::CustomError;
+use crate::http_response::error_handler::CustomError;
 use crate::http_response::HttpCodeW;
 use crate::utils::helpers::{check_if_is_duplicate_key_from_data_base, generate_ic, now_time};
 use sea_orm::QueryFilter;
