+++ src/components/emergency/schedule/emergency_allocation.rs
@@ -4,7 +4,7 @@ use sea_orm::*;
 
 use crate::entity::sea_orm_active_enums::{AmbulanceStatusEnum, EmergencyStatusEnum};
 use crate::entity::{ambulance, emergency};
-use crate::error_handler::CustomError;
+use crate::http_response::error_handler::CustomError;
 use crate::http_response::HttpCodeW;
 use crate::utils::helpers::{calculate_distance, now_time};
