+++ src/components/hospital/services.rs
@@ -1,7 +1,7 @@
 use crate::entity::hospital::{ActiveModel, Column, Entity, HospitalRequestBody, Model};
 use chrono::NaiveDateTime;
 
-use crate::error_handler::CustomError;
+use crate::http_response::error_handler::CustomError;
 use crate::http_response::HttpCodeW;
 use crate::shared::{PaginatedResponse, PaginationInfo};
 use crate::utils::helpers::{check_if_is_duplicate_key_from_data_base, generate_ic, now_time};
