+++ src/components/person/routes.rs
@@ -1,6 +1,6 @@
 use crate::components::person::PersonService;
 use crate::entity::person::PersonRequestBody;
-use crate::error_handler::CustomError;
+use crate::http_response::error_handler::CustomError;
 use crate::http_response::http_response_builder;
 use actix_web::{HttpResponse, get, post, web};
 use sea_orm::DatabaseConnection;
