+++ src/components/staff/routes.rs
@@ -1,6 +1,6 @@
 use super::services::StaffService;
 use crate::entity::staff::StaffRequestBody;
-use crate::error_handler::CustomError;
+use crate::http_response::error_handler::CustomError;
 use crate::http_response::http_response_builder;
 use actix_web::{HttpResponse, post, web};
 use sea_orm::DatabaseConnection;
