+++ src/components/staff/services.rs
@@ -7,7 +7,7 @@ use crate::entity::sea_orm_active_enums::StaffRoleEnum;
 use crate::entity::staff::{
     ActiveModel, Column, Entity, Model, Relation, StaffRequestBody, StaffWithPerson,
 };
-use crate::error_handler::CustomError;
+use crate::http_response::error_handler::CustomError;
 use crate::http_response::HttpCodeW;
 use crate::utils::helpers::{check_if_is_duplicate_key_from_data_base, generate_ic};
 use chrono::{Local, NaiveDateTime};
