+++ src/http_response/mod.rs
@@ -2,6 +2,7 @@ mod create_response;
 mod http_code_w;
 pub(crate) mod http_response_builder;
 mod response_object;
+pub mod error_handler;
 
 pub use create_response::*;
 pub use http_code_w::*;
