+++ src/main.rs
@@ -3,7 +3,7 @@ use crate::open_api::init;
 use actix_cors::Cors;
 use actix_web::http::header;
 use actix_web::middleware::Logger;
-use actix_web::{App, HttpServer, web};
+use actix_web::{web, App, HttpServer};
 use chrono::Local;
 use dotenv::dotenv;
 use env_logger::{Builder, Env};
@@ -15,7 +15,6 @@ use utoipa_swagger_ui::SwaggerUi;
 mod components;
 mod db;
 mod entity;
-mod error_handler;
 mod http_response;
 mod open_api;
 mod shared;
